+++ app/src/main/java/ru/practicum/android/diploma/filter/ui/FilterCountryFragment.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma
+package ru.practicum.android.diploma.filter.ui
 
 import androidx.fragment.app.Fragment
