+++ app/src/main/java/ru/practicum/android/diploma/filter/ui/FilterPlaceToWorkFragment.kt
@@ -1,12 +1,13 @@
-package ru.practicum.android.diploma
+package ru.practicum.android.diploma.filter.ui
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import ru.practicum.android.diploma.R
 
-class JobDetailsFragment : Fragment() {
+class FilterPlaceToWorkFragment : Fragment() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
     }
@@ -17,6 +18,6 @@ class JobDetailsFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_job_details, container, false)
+        return inflater.inflate(R.layout.fragment_filter_place_to_work, container, false)
     }
 }
