+++ app/src/main/java/ru/practicum/android/diploma/team/ui/DevTeamFragment.kt
@@ -1,10 +1,11 @@
-package ru.practicum.android.diploma
+package ru.practicum.android.diploma.team.ui
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import ru.practicum.android.diploma.R
 
 class DevTeamFragment : Fragment() {
     override fun onCreate(savedInstanceState: Bundle?) {
