+++ app/src/main/res/layout/fragment_dev_team.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DevTeamFragment">
+    tools:context=".team.ui.DevTeamFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
