+++ app/src/main/res/layout/fragment_favorites_jobs.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FavoritesJobsFragment">
+    tools:context=".favourites.ui.FavoritesJobsFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
