+++ app/src/main/res/layout/fragment_filter_place_to_work.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FilterPlaceToWorkFragment">
+    tools:context=".filter.ui.FilterPlaceToWorkFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
