+++ app/src/main/res/layout/fragment_filter_region.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FilterRegionFragment">
+    tools:context=".filter.ui.FilterRegionFragment">
 
     <TextView
         android:layout_width="match_parent"
