+++ app/src/main/res/layout/fragment_filter_settings.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:gravity="center"
-    tools:context=".FilterSettingsFragment">
+    tools:context=".filter.ui.FilterSettingsFragment">
     <TextView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
