+++ app/src/main/res/layout/fragment_job_details.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".JobDetailsFragment">
+    tools:context=".details.ui.JobDetailsFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
