+++ app/src/main/res/navigation/navigation_graph.xml
@@ -8,7 +8,7 @@
 
     <fragment
         android:id="@+id/searchJobFragment"
-        android:name="ru.practicum.android.diploma.SearchJobFragment"
+        android:name="ru.practicum.android.diploma.search.ui.SearchJobFragment"
         android:label="fragment_search_job"
         tools:layout="@layout/fragment_search_job">
         <action
@@ -20,22 +20,22 @@
     </fragment>
     <fragment
         android:id="@+id/jobDetailsFragment"
-        android:name="ru.practicum.android.diploma.JobDetailsFragment"
+        android:name="ru.practicum.android.diploma.details.ui.JobDetailsFragment"
         android:label="fragment_job_details"
         tools:layout="@layout/fragment_job_details" />
     <fragment
         android:id="@+id/favoritesJobsFragment"
-        android:name="ru.practicum.android.diploma.FavoritesJobsFragment"
+        android:name="ru.practicum.android.diploma.favourites.ui.FavoritesJobsFragment"
         android:label="fragment_favorites_jobs"
         tools:layout="@layout/fragment_favorites_jobs" />
     <fragment
         android:id="@+id/devTeamFragment"
-        android:name="ru.practicum.android.diploma.DevTeamFragment"
+        android:name="ru.practicum.android.diploma.team.ui.DevTeamFragment"
         android:label="fragment_dev_team"
         tools:layout="@layout/fragment_dev_team" />
     <fragment
         android:id="@+id/filterSettingsFragment"
-        android:name="ru.practicum.android.diploma.FilterSettingsFragment"
+        android:name="ru.practicum.android.diploma.filter.ui.FilterSettingsFragment"
         android:label="fragment_filter_settings"
         tools:layout="@layout/fragment_filter_settings">
         <action
@@ -53,23 +53,23 @@
     </fragment>
     <fragment
         android:id="@+id/filterPlaceToWorkFragment"
-        android:name="ru.practicum.android.diploma.FilterPlaceToWorkFragment"
+        android:name="ru.practicum.android.diploma.filter.ui.FilterPlaceToWorkFragment"
         android:label="fragment_filter_place_to_work"
         tools:layout="@layout/fragment_filter_place_to_work" />
     <fragment
         android:id="@+id/filterRegionFragment"
-        android:name="ru.practicum.android.diploma.FilterRegionFragment"
+        android:name="ru.practicum.android.diploma.filter.ui.FilterRegionFragment"
         android:label="fragment_filter_region"
         tools:layout="@layout/fragment_filter_region" />
 
     <fragment
         android:id="@+id/filterDepartmentFragment"
-        android:name="ru.practicum.android.diploma.FilterDepartmentFragment"
+        android:name="ru.practicum.android.diploma.filter.ui.FilterDepartmentFragment"
         android:label="fragment_filter_department" />
 
     <fragment
         android:id="@+id/filterCountryFragment"
-        android:name="ru.practicum.android.diploma.FilterCountryFragment"
+        android:name="ru.practicum.android.diploma.filter.ui.FilterCountryFragment"
         android:label="fragment_filter_country" />
 
 </navigation>
